refactor: reorganized the compute module
