+++ src/compute/ShaderManager.py
@@ -1,10 +1,8 @@
 import functools
-import os
 import glfw
 import OpenGL.GL as gl
 import numpy as np
 
-from PIL import Image
 from typing import Callable, TypeVar, ParamSpec
 
 P = ParamSpec("P")
@@ -201,144 +199,3 @@ class ShaderManager:
                 )
 
         return final_output
-
-
-@ShaderManager.shader_context(
-    "grid_to_tex", os.path.realpath("src/compute/shaders/grid_to_tex.glsl")
-)
-def write_grid_to_texture(
-    interpolated_grid: np.ndarray, /, shader: str, output_filename="image.png"
-):
-    """
-    Writes a 2D interpolated grid to a texture using an OpenGL compute shader, then saves it as a PNG image.
-
-    Args:
-        interpolated_grid (np.ndarray): The 2D grid data to be saved.
-        shader (str): The compute shader program used to write data to the texture.
-        output_filename (str, optional): Name of the output PNG file. Defaults to "image.png".
-
-    Raises:
-        ValueError: If the input grid is not 2D.
-    """
-    interpolated_grid = np.array(interpolated_grid, dtype=np.float32)
-    grid_height, grid_width = interpolated_grid.shape
-
-    # Create OpenGL texture
-    texture = gl.glGenTextures(1)
-    gl.glBindTexture(gl.GL_TEXTURE_2D, texture)
-    gl.glTexStorage2D(gl.GL_TEXTURE_2D, 1, gl.GL_RGBA32F, grid_width, grid_height)
-
-    # Bind texture to the shader
-    gl.glBindImageTexture(
-        1, texture, 0, gl.GL_FALSE, 0, gl.GL_WRITE_ONLY, gl.GL_RGBA32F
-    )
-
-    # Create OpenGL buffer for the grid data
-    interpolated_data = interpolated_grid.flatten()
-    buffer = gl.glGenBuffers(1)
-    gl.glBindBuffer(gl.GL_SHADER_STORAGE_BUFFER, buffer)
-    gl.glBufferData(
-        gl.GL_SHADER_STORAGE_BUFFER,
-        interpolated_data.nbytes,
-        interpolated_data,
-        gl.GL_STATIC_DRAW,
-    )
-    gl.glBindBufferBase(gl.GL_SHADER_STORAGE_BUFFER, 0, buffer)
-
-    # Set shader uniforms
-    grid_width_loc = gl.glGetUniformLocation(shader, "gridWidth")
-    grid_height_loc = gl.glGetUniformLocation(shader, "gridHeight")
-    gl.glUniform1i(grid_width_loc, grid_width)
-    gl.glUniform1i(grid_height_loc, grid_height)
-
-    # Dispatch the compute shader
-    workgroup_size = 16
-    gl.glDispatchCompute(
-        (grid_width + workgroup_size - 1) // workgroup_size,
-        (grid_height + workgroup_size - 1) // workgroup_size,
-        1,
-    )
-    gl.glMemoryBarrier(gl.GL_SHADER_IMAGE_ACCESS_BARRIER_BIT)
-
-    # Retrieve texture data
-    output_data = np.empty((grid_height, grid_width, 4), dtype=np.float32)
-    gl.glGetTexImage(gl.GL_TEXTURE_2D, 0, gl.GL_RGBA, gl.GL_FLOAT, output_data)
-
-    # Convert to grayscale image and save as PNG
-    output_image = (output_data[:, :, 0] * 255).astype(np.uint8)
-    real_path = os.path.realpath(output_filename)
-    Image.fromarray(output_image).save(real_path)
-
-    # Cleanup
-    gl.glDeleteTextures(1, [texture])
-    gl.glDeleteBuffers(1, [buffer])
-
-    print(f"Saved interpolated grid to {real_path}")
-
-
-@ShaderManager.shader_context(
-    "cubic_spline", os.path.realpath("src/compute/shaders/cubic_spline.glsl")
-)
-def cubic_spline(data: np.ndarray, samples: int, *, shader: str) -> np.ndarray:
-    """
-    Performs cubic spline interpolation on a 1D array using an OpenGL compute shader.
-
-    Args:
-        data (np.ndarray): 1D array of input values to be interpolated.
-        samples (int): Number of interpolated samples between adjacent data points.
-        shader (str): The compute shader program used for interpolation.
-
-    Returns:
-        np.ndarray: The interpolated 1D array.
-
-    Raises:
-        ValueError: If the input data is not a 1D array.
-    """
-    data_size = len(data)
-    output_size = data_size * samples + (data_size - samples)
-
-    input_buffer = np.array(data, dtype=np.float32)
-    output_buffer = np.zeros(output_size, dtype=np.float32)
-
-    gl.glUseProgram(shader)
-
-    # Create buffers and bind them to shader storage
-    input_ssbo = gl.glGenBuffers(1)
-    gl.glBindBuffer(gl.GL_SHADER_STORAGE_BUFFER, input_ssbo)
-    gl.glBufferData(
-        gl.GL_SHADER_STORAGE_BUFFER,
-        input_buffer.nbytes,
-        input_buffer,
-        gl.GL_STATIC_DRAW,
-    )
-    gl.glBindBufferBase(gl.GL_SHADER_STORAGE_BUFFER, 0, input_ssbo)
-
-    output_ssbo = gl.glGenBuffers(1)
-    gl.glBindBuffer(gl.GL_SHADER_STORAGE_BUFFER, output_ssbo)
-    gl.glBufferData(
-        gl.GL_SHADER_STORAGE_BUFFER,
-        output_buffer.nbytes,
-        output_buffer,
-        gl.GL_DYNAMIC_COPY,
-    )
-    gl.glBindBufferBase(gl.GL_SHADER_STORAGE_BUFFER, 1, output_ssbo)
-
-    # Set shader uniforms
-    gl.glUniform1i(gl.glGetUniformLocation(shader, "samples"), samples)
-    gl.glUniform1i(gl.glGetUniformLocation(shader, "dataSize"), data_size)
-
-    # Run compute shader
-    gl.glDispatchCompute((output_size + 255) // 256, 1, 1)
-    gl.glMemoryBarrier(gl.GL_SHADER_STORAGE_BARRIER_BIT)
-
-    # Retrieve results from output buffer
-    gl.glBindBuffer(gl.GL_SHADER_STORAGE_BUFFER, output_ssbo)
-    output_data = np.frombuffer(
-        gl.glGetBufferSubData(gl.GL_SHADER_STORAGE_BUFFER, 0, output_buffer.nbytes),
-        dtype=np.float32,
-    )
-
-    # Cleanup
-    gl.glDeleteBuffers(2, [input_ssbo, output_ssbo])
-
-    return output_data
